feat: moved components for reusage
