+++ core/ui-common/src/main/kotlin/com/wire/android/ui/common/bottomsheet/WireBottomSheetDefaults.kt
